Core/Misc: Turn ChrSpecialization into enum
Port From (https://github.com/TrinityCore/TrinityCore/commit/98007f859b7318570c0c923a00aa32fc485c8ec8)

+++ Source/Game/Entities/Item/ItemTemplate.cs
@@ -241,7 +241,7 @@ namespace Game.Entities
 
             uint spec = player.GetLootSpecId();
             if (spec == 0)
-                spec = player.GetPrimarySpecialization();
+                spec = (uint)player.GetPrimarySpecialization();
             if (spec == 0)
                 spec = player.GetDefaultSpecId();
